requirements.txt
# Core data stack
pandas
numpy
scipy
scikit-learn
pyyaml
openpyxl
tabulate

# Visualization
matplotlib
seaborn
plotly

# Reporting / Jupyter
jupyter

# Report generation
pypandoc
weasyprint   # for PDF export (no LaTeX needed)


pyppeteer 
markdown